Move the classic .interact() method into the standalone interact.py, rename as interact_classic, add kwarg to new interact.
